refactor(core/search): Rename searchResultsType.registry
