+++ app/scripts/modules/core/src/search/searchResult/searchResultsType.registry.ts
